+++ moe/optimal_learning/python/repeated_domain.py
@@ -79,6 +79,22 @@ class RepeatedDomain(DomainInterface):
         """Return a list of ClosedIntervals representing a bounding box for this domain."""
         return self._domain.get_bounding_box()
 
+    def get_constraint_list(self):
+        """Return a list of lambda functions expressing the domain bounds as linear constraints. Used by COBYLA.
+
+        Calls ``self._domain.get_constraint_list()`` for each repeat, writing the results sequentially.
+        So output[0:2*dim] is from the first repeated domain, output[2*dim:4*dim] is from the second, etc.
+
+        :return: a list of lambda functions corresponding to constraints
+        :rtype: array of lambda functions with shape (num_repeats * dim * 2)
+
+        """
+        constraints = []
+        for i in xrange(self.num_repeats):
+            # Using start_index, start each domain at the correct index when flattening out points in COBYLA.
+            constraints.extend(self._domain.get_constraint_list(start_index=self.dim * i))
+        return constraints
+
     def generate_random_point_in_domain(self, random_source=None):
         """Generate ``point`` uniformly at random such that ``self.check_point_inside(point)`` is True.