+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * Added startup message to REST server including tips for OSX users (#400)
   * Added GPU support to ``cpp_wrappers.expected_improvement.multistart_expected_improvement_optimization``; requires ``max_num_threads == 1`` until future multi-GPU support (#368)
+  * Added the COBYLA optimizer to the expected improvement optimization class. (#370)
 
 * Changes